+++ TexolBilling/TexolBilling/BAL/SalesDetails.cs
@@ -33,7 +33,7 @@ namespace TexolBilling.BAL
 
         }
         
-        public int InsertSalesItem(string SaleTrNo,int ItemId,int rate,int quantity)
+        public int InsertSalesItem(string SaleTrNo,int ItemId,decimal rate,int quantity)
         {
             string sql= "insert into salesitem_tbl (SalesTransactionNo,ItemId,Rate,Quantity) values('"+SaleTrNo+"',"+ItemId+","+rate+","+quantity+")";
             int i = db.InsertData(sql);
@@ -44,7 +44,7 @@ namespace TexolBilling.BAL
             string sql = "select  item_tbl.ItemName,salesitem_tbl.Rate,salesitem_tbl.Quantity,(salesitem_tbl.Rate*salesitem_tbl.Quantity) as TotalAmount from item_tbl inner join salesitem_tbl on salesitem_tbl.ItemId=item_tbl.ItemId where salesitem_tbl.SalesTransactionNo='" + TransactionNo + "'";
             return db.GetData(sql);
         }
-        public int InsertDataToSalesTbl(string salesTrno, DateTime date, int customerid, int tax, int tamount,int subtotal)
+        public int InsertDataToSalesTbl(string salesTrno, DateTime date, int customerid, decimal tax, decimal tamount,decimal subtotal)
         {
             string sql = "insert into sales_tbl(SalesTransactionNo,SalesDate,CustomerId,TaxAmount,TotalAmount,SubTotal) values('" + salesTrno + "','" + date + "'," + customerid + "," + tax + "," + tamount + ","+subtotal+") ";
             int i = db.InsertData(sql);
